Fix race condition between initial scheduler await
- This commit fixes a bug where the child process and the
scheduler would race for the first transition in the
schedule, causing the invariant that only one of the processes
is running to be invalidated. This ultimately caused a data race
in the shared memory region (deferencing garbage).

The solution is for the intial thread_await called by the child
process to *exclusively* wait for the parent process to signal it.
Implicitly, `dpor_run` and `thread_await_dpor_scheduler()` assume
that the other process is asleep waiting on the binary semaphore; the
new method does not make such an assumption (it assumes the other
process is awake)