Add potential fix to hashtable
- The current implementation of the hash table uses
linear probing. Upon removing elements from the hash
table, however, linear probing demands that the
entries following the entry be removed be inspected
to prevent hash-collided keys from improperly not
being found in the future. The solution is to continue
searching the hash table for replacements until you
hit an emoty entry