Move visible-object state determination to scheduler
- This commit moves the determination of the state of
visible operations to the scheduler parent process. This
is needed because there is no reliable way for the child process
to know the state of everything unless the concurrent system
resides in shared memory. This however won't work for some of the
dynamic types such as the hash_table and hash_set which malloc
into their own private address spaces

The solution is simply to push state determination to the scheduler,
which already knows the state of all visible objects anyway. This
indeed fits in better with detecting undefined behavior as well,
since the scheduler is the one that would be able to determine that in
the context of other transitions