Resolve infinite loop in hash_table and add pthread_create support
This commit adds support for the pthread_create wrapper operation
while backtracking.

What remains is to add

1. pthread_join into the mix, among other thread operations
2. ensure that the backtracking is working as expected with
multiple threads

The second point will be investigated futher in subsequent commits

Secondly, a bug in the hash_table implementation was patched
whereby a hash_table that was completely full would enter an
infinite loop attempting to get an element that is *not* already
in the table. The solution is simply to keep track of whether we
have searched the entire table (by comparing how many elements are
in the table to the number we have searched through)