Rework backtracking logic and resolve IPC deadlock
- This commit makes some small modifications to the
backtracking logic.

A deadlock was identified whereby the scheduler would invoke
a thread that had not yet been created in the child process
to be scheduled to run. The bug is resolved by marking thread
creation *not* co-enabled with the THREAD_START operation. This
example illustrates a central problem that will hopefully be
addressed shortly, viz. how do we deal with thread/resource creation
events?