Remove iterative deepening and search continuously
- This commit removes the iterative deepening aspect
of the current implementation of the algorithm, as it does
not technically work out in the way that DPOR was designed.
The issue is that by restricting ourselves to a certain depth,
we are not properly updating the backtrack sets and are effectively
"starving" new threads created in the source program with
the scheduler.

The last remaining bug is to have the t_next stack have a special
transition type with *reference* semantics as pointed by Luka
in a previous commit. However, we need to define a new struct,
as value semantics are essential to the implementation of the
backtracking portion of DPOR; i.e., we need both value and
reference semantics to properly implement this algorithm.

Once this bug is resolved, I am highly confident the algorithm
will be working correctly with mutexes and thread creation.
Almost there! There is still a lot of work that needs to be done,
which will be highlighted in GH issues to come