Add brief comment describing the scheduler race condition
- There was a race condition in the wrapper functions
called into by the simulated program related to pthread_create.
It was possible that a thread was created and, before it had a
chance to register itself in the csystem in its own address space,
another thread was scheduled to be created. This resulted in a
race between the two new threads in the child process in terms
of how they were created. The solution is to atomically create
a new thread (wait until the new thread registers itself before
waking the scheduler)