Remove soft registration of visible objects
- This commits removes the ability to "soft-register" a
visible object in DPOR. This would cause aliasing bugs
among visible objects that would potentially cause all
sorts of nasty behavior within the schedler. The solution
is to simply create the objects when they are encountered,
and to persist those objects across the lifetime of the scheduler.

This demands that shared memory handlers know how to handle the
case the visible objects that they encounter are already apart of
the scheduler; i.e., the handlers must check that the system identity
has not already been marked by the scheduler before creating more
objects