Prevent detecting deadlock for traces greater than search depth
- This commit adds an additional check to the deadlock condition
that ensures that only deadlocks that occurred within N visible
operations that count against the search depth will be reported.
It is necessary to add this check as DPOR may decide to backtrack
after more than N transitions have been run and encounter a deadlock
within a trace of more than N transitions. It is still necessary
to perform this work as there is no guarantee that a subsequent
transition by a thread, perhaps only brought to light due to
backtracking, would be excluded from a deadlock of at most N
transitions