Add fflush to output to prevent printf races
There was a bug where trace print-outs were
occurring multiple times. This was (very likely)
due to the fact that writes are buffered by
printf and friends; thus before the print out could
happen in the parent process, it fork()'ed and the
child also got the buffer to print out. This is a
"benign" race in the sense that the actual underlying
values of the classes etc inside the parent process
are correct; but the textual reflection of those values
was incorrectly duplicated

The solution is to flush stdout and stderr in appropriate
spots when printing. Large collections of prints are
flushed together, while prints that are more independent
cause a flush to happen immediately