Add more clever backtracking logic
This commit checks whether a set of candidate
backtracking threads (computed in the pseudocode
in the original Flanagan/Godefroid paper) has
a nonempty intersection with the threads a particular
state stack item is already backtracking on. In such
a scenario, we don't need to bracktrack: we can simply
decide to chose a thread that is already going to be
backtracked on to get an equivalent result (since we
can pick _any_ candidate thread when we have a
non-empty set of them). This could improve performance
in larger programs