Fix GDB trace tracking bug
The bug was resolved by using an _Exit() call
instead of an exit() call; the former doesn't invoke
atexit-handlers while the latter does. In child processes
forked from the scheduler process, we explicitly register
an atexit-handler to "catch" the main thread when it exits
from the main routine to keep the child process alive.

When we want to *actually* terminate the child process (e.g.
when something abnormal happens in the child process), we need
to be mindful of the atexit-handler we previously registered.
Simply calling _Exit() (or _exit()) works here

Subsequent commits will improve exiting behavior by always
choosing the correct exit function based on the context (i.e.
whether it is being invoked in the scheduler or a forked child)