Use thread ids exclusively for sleep sets
We take advantage of an observation that
transitions in any sleep set

1. must each be run by different threads; and
2. for any thread `p`, if there is a transition
run by `p` in the sleep set, then it MUST be
the *next* transition `p` will run

The former follows from the fact that two transitions
run by the same thread are dependent. The latter is
a bit trickier to observe but follows from the former.

Observe that if a transition is in the sleep set at state
S, it was either a) run from that state, or it was both
a) added at an earlier state S' and b) did not execute
at any point between S and S'. In the first case,
we're done: since `p` will not execute from S but is contained
in the sleep set, whatever it would have run next is
now in the sleep set. The second follows from the first:
since no other transition by `p` could have been run
to reach state S, it must still be the next transition `p`
ran when it was added to S'