Add subtle optimization when backtracking
This commit adds the subtle optimization descibed
on page 6 of Flannagan and Godefroid's DPOR
paper. The semantics hopefully match that of what
they describe in the paper (viz. performing another
persistent set computation when we encounter
the same thread to backtrack on). We attempt to
avoid the persistent set re-computation whenever
we can select a different thread