Add build scripts
This commit adds build scripts to the
'scripts' directory which automatically
invoke cmake in the appropriate way to
compile the project. The scripts should be
run using the `zsh` shell. Bash would probably
also work in this case, though we should add
in bash scripts as well in case of any differences
between the two shells