Add state reversal index stack
This commit introduces a new MCSortedStack
templated type which is used to keep track
of the indices into the state stack at
which the transition bringing the concurrent system
into that state is irreversible from that state.
This information can be used to quickly determine
whether or not a sequence of transitions can be reverted.
Currently, no transition is implemented for state
reversal

BUG-FIX:

  Indicies into the transition stack were
  improperly mixed with indices into the
  state stack, causing the data held by
  the `irreversableStatesStack` of MCState
  to incorrectly persist data across traces.

  The solution is simply to ensure that
  only state stack indices are added to the
  irreversibleStatesStack.