Move common transition functions to MCTransition
Functions that were previously static members
of the MCTransitionFactory are static member
functions of the MCTransition class. It makes
more sense for those functions to remain
a part of the MCTransition interface since "factory"
typically implies that transitions are _created_
as opposed to _compared_.

Subsequent commits will introduce a new method
on transitions that will by default check if the
thread is enabled before asking the transition if
it's enabled.