Add .clang-format description to McMini
This commit adds a .clang-format configuration
file that is used by the `clang-format` program
to enforce code styling automatically. Clang-format
usually can be installed with your favorite package
manager (e.g. apt on Ubuntu). If you run clang-format
on a file from the top-level directory with the `-i`
flag, clang-format will use the C++ configuration in
the clang-format file to adjust any C++ source code
in the file specified.