Add glibc group1-group2 signal policy
This commit introduces the group1-group2 signal
policy implemented by glibc. Notice that technically
the full specification of a condition variable requires
that a wakeup policy also be defined; reading through
the source code of nptl indicates that a FIFO wakeup
policy is implemented by glibc. Thus, to emulate
condition variables as implemented in glibc, you
should specify a wakeup policy of fifo and a signal
policy of group1-group2.