Add enqueue operation to reader-writer locks
An extra enqueue operation is needed in order
to gain control over the ordering in which
threads acquire the lock. This appears to be a common
theme amongst transitions that McMini needs to
control the thread scheduling therefrom. It turns
out that McMini is more of an emulator as opposed
to a "transparent spy". This will prove interesting
as we continue to develop McMini, where it will
eventually take a more transparent role