Fix reader-writer lock state transfer bug
Effectively, the copy() method of the MCRWLock
was improperly implemented and didn't initialize
the `type` field to be that of the copied lock.
Thus when McMini later regenerates object states,
it incorrectly transitioned a lock from one type
(say writer preferred) to another type. Since the
type was uninitialized, this was invoking undefined behavior.
Since the C++ new() implementation likely zeros its
allocated memory before initialization, this meant the
lock would suddenly shift to behaving as a writer-preferred
lock (since, as a raw value based on the declaration of the
MCRWLock::Type enum, writer_preferred is declared first).
This behavior matches the erroneous traces. Simply
correcting the mistake in copying object states across
objects resolves the issue