Disable state reversal for MCThreadStart/Create
There is a logic error in MCThreadStart. A
transition is correctly disabled when the thread
is either an embryo or dead. However, reversing
a thread state transition should not bring the
transition back into the embryo state; rather
the thread should probably remain unaffected,
leaving the despawning to MCThreadCreate. We need
to think about this a bit more before allowing
either to take part in state reversal.