Add tests for double-writer locks
This commit introduces two new test programs:
reader_writer_rwwlock.c and reader_two_writers_cond.c.
The latter is a straightforward extension of the
condition variable implemention of reader/writer
locks with only a single writer type. The latter type
requires some special care when building though.

Since the pthread_rwwlock_t primitive and its
operations don't really exist, we need to create them
and package them into a dynamic library against which
a test program can link. Creating the dynamic library
isn't difficult: we follow the standard build pattern
as described below:

  gcc -c -fpic rwwlock.c
  gcc -shared -o librwwlock.so rwwlock.o

The trouble comes when linking the test program against
the library. Clearly, since this build library is not
in the standard library paths searched for dynamic
libraries, although the program compiles successfully,
it will not execute properly at runtime. Only at runtime
are the dynamic library links searched. If you inspect
the executable using e.g. _ldd_, you'll see that the path
isn't found.

The simplest solution is to add $ORIGIN to the runpath
of the executable when linking. Viz., we need to create
the test program reader_writer_rwwlock.c as follows:

  gcc -g3 -O0 -Wl,-rpath,"$ORIGIN" reader_writer_rwwlock.c \
   -o reader_writer_rwwlock -pthread -L. -lrwwlock

Note that we assume that librwwlock has already been built.
This will allow the program to execute successfully so
long as the compiled library is in the same directory
as the executable. See ld.so(8) for more details
on $ORIGIN