Add quick hack to support assertion failure traces
This commit introduces a quick hack whereby the child
process signals the parent (scheduler) process in the
event that child encounters an assertion failure. It leverages
the raise(2) and kill(2) system calls and SIGUSR2 signal.
Later, the scheduler should be more robust and handle
all exits of the trace process