Replace is_trivially_copyable check with macro
The is_trivially_copyable<T> templated struct is
not available on GCC prior to version 5.0. Testing
for whether the struct is trivially copyable is
pretty crucial as it determines whether or not the
struct MCSharedTransition can be copied into shared
memory (which is essential for McMini to function
properly).

How the check itself takes place is rather unimportant:
it's a static_assert that will fail at compile-time if
the MCSharedTransition struct is changed in such as way
as to prevent it from being trivially copyable.

The solution is to use the __has_trivial_copy() type
trait available on older versions of GCC (including
the GCC 4.8 which is the version Khoury uses)