Make mc_do_model_checking() more readable
This commit is a pretty good step forward to making
mc_do_model_checking() more readble. There are still
some more adjustments that could be made to increase
the readability; but larger gains will be made by increasing
the readability/understanadbility of the other C functions
that play a role in the data path of McMini, viz.
mc_search_dpor_branch_following_thread() and others.