Add GDB hook for interactive debugging
This commit introduces the hook for GDB
so that it can provide interactive
debugging with a target program. The hook
is a simple `while`-loop that repeatedly
executes the current trace until GDB decides to
tell McMini to continue its search through
the state space as usual.

There are two global variables that GDB manipulates
in order to control the behavior of McMini:

1. The `gdbInterruptTraceId` is used to mark a
particular trace for repeated re-execution under
the control of the user. McMini uses this value
to determine whether it should abort re-execution
of the current trace and continue searching
the state space as usual. McMini will simply pick
up where it left off before GDB took control of
the particular trace. Note that GDB can make McMini
re-execute a trace an arbitrary number of times
without affecting future state searches

2. The boolean value `rerunCurrentTraceForGDB` can
be set to `true` in the middle of the re-execution
of a trace to signal to McMini that the trace should
be *re*-executed. McMini will abandon the current trace
and will start completely anew (i.e. a completely new
process will be created).

The second global is necessary as there is a distinction
between needing to re-execute the current trace and
finishing debugging the current trace.

Note that it is still GDB's responsibility to accurately
control the number of times that McMini schedules threads
for execution. For example, if GDB wishes to bring the
user 6 transitions into the current trace, it would need
to allow McMini to schedule 6 threads for execution.