Add total transition depth check for McMini
A simple check was added to McMini that prevents
execution beyond the maximum number of transition
McMini can handle in any given trace. Reducing the
number of transitions each thread can execute can help
reduce the total number of transitions in the trace.

Note, however, McMini may need to schedule multiple
threads to their thread execution depths in order to
get a full view of the possible deadlocks within
a given number of steps (for correctness). Thus
limiting the execution depth may not be sufficient
to limit the total trace depth to a reasonable
number (consider if the program spawns several
threads)