Add dedicated signal handling functions
The signal handling registration logic has
been improved in two ways:

1. The logic is moved in the mcmini/signals.h
header file out from mcmini_private.cpp

2. sigaction() and friends are used in place
of signal() to allow sem_wait() to be interrupted
as we expect

SIGINT is now also explicitly captured and noted
by McMini. The shared memory is cleaned up in all
cases where McMini exits early.

NOTE: We should note that the functions we are
calling inside the signal handlers are probably not
async-signal safe. This is probably an issue and
we should make note of it in the future