Add support for reacting to SIGCHLD
When a trace process exits, unexpectedly
or otherwise, a SIGCHLD is delivered to
the scheduler process. To determine the
difference between a failed trace process
and a trace process exiting because it was
sent a SIGUSR1, the siginfo_t struct passed
to the signal handler registered by sigaction(2)
is inspected for its sa_code field to determine
if the child exited cleanly (i.e. if sa_code is
CLD_EXITED according to the man pages).

In the future, since the functions currently
called within the signal handler are not async-signal
safe, we'll need to set a global sigatomic_t value
and check it in the appopriate locations instead of
exiting directly