Add UB check for pthread_cond_wait() with two locks
Invoking pthread_cond_wait() with the same condition
variable and two different mutexes is undefined behavior.
McMini will now detect such scenarios as such and report
them as in any other trace.

This commit also resolves a SEGFAULT in McMini when
it attempted to determine whether or not a MCCondEnqueue
were dependent with other transitions in the transition
stack. The bug essentially could be boiled down to the
fact that McMini attempts to compute clock vectors
before running the transitions themselves. Simulation
of the transition moves the checker side into its NEXT
representation of the state of the program it is checking,
while clock vectors need to be computed for the prior state.

To resolve the issue, MCCondEnqueue should not ask a
question about on the mutex "associated with" the condition
variable and instead should ask whether or not other transitions
are dependent/enabled with the mutex that it owns (and
later passes to the condition variable after execution).