Prepare transition to `ConditionVariablePolicy`
The bug that was discovered in McMini with
respect to condition variables opened up more
discussion around supporting condition variables
in general. To allow users to completely define
the semantics of their transitions, the complicated
semantics of the different types of condition
variables supported by McMini are delegated to the
MCConditionVariable object and friends themselves.

That said, the actual implementation of the condition
variable to match the semantics of pthread_cond_wait()
etc. is going to be adjusted to account for the
atomicity of a pthread_cond_signal() and
pthread_cond_broadcast(). This atomicity was subtlely
different prior to the changes that will come. Those
changes in addition to the actual bug fix will be
added in subsequent commits