Remove threads from WakeGroup even if broadcast
We want to remove a thread from the WakeGroup
even if it is broadcast. Prior, we only removed
the thread from other wake groups if it consumed
a signal. But we must also do so if it "consumes"
a broadcast: if the thread re-entered the condition
variable, we would not want to say that the thread
was eligible to wake up since it had already "escaped"
the condition variable at the time the other signals
arrived. This is an implementation detail, but an
important one. Since McMini simulates atomically waking
a thread and its consumption of a signal, this forces
us to ensure that things are ready for a thread which
exits and re-enters a condition variable many times