Deterministically pick threads in the backtrack set
DPOR does not impose any restrictions on which
thread should be picked first from a backtrack
set when continuing its search. Prior to this commit,
McMini arbitrarily selected the first thread it
found contained in an *unordered* collection
of threads. The problem is that the *some*
ordering is ultimately defined by the standard
library implementation of `std::unordered_set<>`,
but this may vary between distributions. It is
well known that DPOR with sleep sets can exhibit
up to an order of magnitude difference in performance
according to the particular thread orderings that
it follows (since those orderings hit sleep-set
blocked traces differently; see the 2017
paper on ODPOR for details).

Since we aim for McMini to be deterministic, we
need to provide our own ordering. Thus we now
arbitrarily pick the thread with the smallest
thread id from the backtracking set before
choosing other threads. This ensures that McMini
will search in the same way regardless of the
underlying implementation of std::unordered_set<>