Add state_view and state_sequence implementations
This commit adds the first implementations for state and state_view.
The `state_view` will be fed to a transition which will attempt to
modify a clone of the view. Modifications to the clone will be forwarded
to the underlying sequence

An alternative strategy will be implemented instead in later commits.
this alternative is for transitions to tell you _differences_ in state
which can be applied if needed. This separates the underlying addition
of states to the sequence from the actual (attempted) execution of the
transition itself (these should be independent