Add changes before reverting templated visible object
The templated visible object interface is a bit finnicky
to say the least. Better yet would be to leverage a
double-dispatch table that could be used to correctly
compare two different types of objects and have transition
registration provide the necessary function pointers. A
static_assert<> could even be provided to force end users
to implement said function