Complete implementation for state_sequence
This commit completes the implementation of the
method left unimplemented in state_sequence.cpp. What
remains is to implement the two constructors on `state_sequence`
which take an initial state from which to copy. This isn't
strictly required and more of a niceity; hence, it is ignored
for future commits