Hotfix for std iostream SEGFAULT.
`libmcmini.so` is loaded even before the
standard C++ library. Therefore, it is unsafe
to use the iostreams provided by the C++ standard
library as they will not be properly initialized
when McMini is running.

A future change should be to move the loading of
the C++ standard library to _after_ libmcmini.so.
This is not currently the case.