Add runner mapping management to state/mutable_state
This commit introduces methods for managing runners
in states. Transitions operate on states and states
alone; but they are themselves executed by runners.
It is useful, and necessary sometimes (e.g. with the
thread_start transition), to identify the modeled states
of these runners in the model via the runner id instead
of the object id directly. Here we note that it is
sensible to add this mapping directly to the state:
the transition that these runners are executing is
outside the scope of the state, but simply tracking which
objects they refer to is sensible.