Record the presence of runners in state_sequence for elements
This commit introduces changes to how the elements
in the state_sequence are informed of runners. Previously,
elements were not informed of changes to the number of
runners and thus were not properly synchronizes with the
changes forwarded to the `state_sequence`. This commit
addresses that issue.

Additionally, the use of the `volatile_mem_stream` has been
removed. It turns out to be needlessly complicated and does
not provide much over what a simply dereference from a raw
pointer doesn't itself already achieve (and is probably much
slower anyway).