Wrap main in try-catch for unhandled exceptions
This commit wraps `main` in a try-catch to ensure that
all exceptions that make it all the way to the top level
do not result in a call to `std::terminate()`. This ensures
that proper destruction of static global variables occurs
even when an exception remains unhandled throughout the entire
call stack (save for main).