Fix subtle issue with `kill(2)` delay to grandchild process
This commit resolves a subtle issue with the way the
architecture changes to McMini affected the plumbing details
in `libmcmini.so`. When `libmcmini.so` is in the "template"
mode, it is signaled by the `McMini` process to create a new
child process. When McMini no longer needs the child process,
it will now send a signal to the child process the template
created previously by calling `fork(2)`. However, signal delivery
is _asynchronous_, which means that exiting from the clean up
in the destructor of `local_linux_process` does _not_ necessarily
mean that the process itself has died, even though the destructor
calls `kill(2)`.

To resolve the issue, we effectively clump together the action of
"forking the template to produce a new child process" with
"waiting for the previous child process to terminate." This is
accomplished with a call to `wait(2)` right before the `sem_wait(2)`
call in `mc_template_process_loop_forever()`. This has the advantage
that the C++ code in the McMini process is already designed to treat
the creation of a new child as asynchronous.