Extract `diff_state` as its own class
This commit introduces `model::diff_state` as a
standalone entity independent of the `state_sequence`
implementation. There is a surprising benefit to doing
this:

The implementation of `transition` is now both more efficient
AND more insightful, the former because we avoid a copy of the
provided state altogether (`diff_state` is cheap in the base case)
AND we return the state that now also encodes the DIFFERENCE
it made on the provided state. This could prove interesting
later