Add prctl(2) call in the McMini and child processes
We use prctl(2) to deliver a `SIGTERM` to the
template process and to the child process in the
event that either the `McMini` process or the parent
process exit unexpectedly.

We should note that this could potentially be risky.
If the template process fails (which would only occur if
`libmcmini.so` has a bug most likely), the child will also
die, but McMini will not be notified of this. The solution
would be to check for the reception of a signal when
sleeping on the semaphores and checking whether an exceptional
condition occured