Remove operator[] from `injective_function`
This resolves a subtle bug with `operator[]`
on the injective_function class. The issue is
that `operator[]` on `std::unordered_map` generates
a value which we then store in the back-mapping, but
the default value that is generated is NOT the value
we want to insert. We opt to use `insert()` instead