Add dedicated signal-handling thread in McMini
This commit adds a dedicated signal handling thread
into the McMini process. This is needed to ensure that
`std::exit()` is called appropriately upon the reception
of a SIGINT signal and ONLY by a single thread. Calling
`std::exit()` ensures that C++ global objects get
destroyed, which is important for McMini's cleanup (in particular
for the shared memory region and dealing with the template
process).

There is a small gotcha though. `pthread_sigmask()` is preserved
across `exec()` boundaries, so we need to clear the mask we
filled at startup in the fork()-ed child process created just
before the template