Add DMTCP no-checkpoing critical sections
This commit adds calls to `dmtcp_enable_ckpt()`
and `dmtcp_disable_ckpt()` in the appropriate places
in the recording code. Care is taken to ensure that
the wrappers properly escape to control of the model
checker at restart-time.

We should note, however, that livelock is still a possibility
as of this commit: it's possible for the user threads to
indefinitely delay checkpointing if they collectively repeatedly
enter and re-enter wrapper functions such that at there is always at
least a single thread recording information.