Refactor fork()/exec() in McMini to `target` class
The (somewhat complicated) code used to spawn a new child process
was refactored into the `real_world::target` class. The code was
previously duplicated between the `dmtcp_process_source` and the
`fork_process_source`.

The motivation is to enable the `dmtcp_process_source` to create
the `dmtcp_coordinator` at construction time and then pass
the `--join-coordinator` flag to `dmtcp_restart` for each new
branch. Furthermore, when we later add support for multithreaded fork
instead of `dmtcp_restart`, the code needed for the McMini process will
be drastically simpler