Shutdown dmtcp_coordinator daemon via dmtcp_command
This commit provides an implementation for enabling multiple
`dmtcp_restart` processes to be used as branches.  The idea is to spawn
a single instance of `dmtcp_coordinator` as a daemon and have subsequent
branches use the `--join-coordinator` flag.

The trouble comes when we want to do clean up and kill the coordinator. To
become a daemon process, the coordinator calls `fork()` and the
parent exits while the child persists.  This means that the true
`dmtcp_coordinator` process in memory is a _grandchild_ of McMini and
not a child process.

To clean up this grandchild, we use `dmtcp_command -q` which tells the
coordinator to shutdown. To prevent conflicts with processes using dmtcp
out-of-the-box (i.e. without any special customization), we point our
`dmtcp_coordinator` daemon to the special port 7778. The `dmtcp_restart`
and `dmtcp_command` processes are likewise pointed to this port.