Temporarily disable prctl(2) for child processes
To clean up the `dmtcp_coordinator` daemon on shutdown, the `mcmini`
process invokes `dmtcp_command -q`. To prevent this from happening,
we disable `prctl` temporarily.

`prctl` is used to aide in the clean up process.  If `mcmini` exits,
whether expected or not, any branches and/or template processes should
be killed.  In the case of `dmtcp_command`, however, we want the process
to outlive `mcmini`.