Fix bug with clock vectors in experimental McMini
This commit resolves a bug with clock vectors that occurred during the
translation from the primary McMini code. The bug occurred because we do
_not_ re-execute transitions against the McMini model when regenerating
previous states; instead, we simply fetch "slices" of the current
sequence of states which "performs" the regeneration for us. The idea is
that states we've already encountered don't need to be regenerated. We
sacrifice the small amount of space needed to store a single trace of
states to improve performance on top of McMini.

After regenerating the model, the model checker must update its own
bookkeeping structures whenever it wants to explore a different part of
the state space. In classic McMini, the coupling between the model
checker and the model meant that the regeneration of the model checking
structures happened for free. However, the model and the model checker
are _no longer_ coupled (which is a win), which means that we need to be
more explicit on the model checking side.

The solution is simply to store the per-thread clock vectors _in each
state_ in the same way that the per-state vector clocks are stored. This
is in contrast to classical McMini, which stores a single per-thread
clock vector that is patches when DPOR tells it to search another branch
of the state space.