Fix move ctor bug with `local_linux_process`
This commit defines the move constructor and move assignment operators
for the class `local_linux_process`. This was hitherto undefined
explicitly, but rather implicitly. That is, the move semantics were
implicitly defined for us. However, the "default" is seldom the right
choice if other operators/assignment overloads are defined (think the
C++ "Rule of 5"). The `local_linux_process` manages a resource
RAII-style, and hence a move ctor/assignment operator are needed.