Add temporary fix for `dmtcp_coordinator` race
This commit patches a bug whereby the `dmtcp_coordinator` daemon spawned
in the ctor of `dmtcp_process_source` has not yet opened its port for
communication before spawning `dmtcp_restart` instances while branching.
The "solution" is to have `dmtcp_coordinator` write the port number to a
file using the `--port-file` flag. We can then wait for this file to be
created and read it to retrieve the correct port number. A busy wait is
used for now, but this can be improved later since it's not super
important for progressing.