Add bare-bones multithreaded_fork_process_source
This commit adds the infrastructure for (but not the implementation of)
the `real_world::process_source` abstraction used from the point of view
of the model checker. Here we base much of the work off of the
template-target branch implementation of `fork_process_source` which we
use for classical model checking.

A new `dmtcp_coordinator` target was added that handles the details of
fetching the port number of the `dmtcp_coordinator` process as well.