Add distinct restart modes for `libmcmini.so`
The `DMTCP_RESTART` `libmcmini_mode` is now split into two cases. The
first case, `DMTCP_RESTART_INTO_BRANCH`, matches the old semantics of
`DMTCP_RESTART`. Here the name makes it clearer that the restarted
process will become a model checking _branch_ process instead of a
template process.

On the other hand, `DMTCP_RESTART_INTO_TEMPLATE` signifies that the
restarted image will become a template process. The difference is that
the userspace threads in the template process must not interact at all
with the model checker.