Fix subtle race between target branch processes
A subtle bug was discovered whereby a race between a
soon-to-be-extinguished target branch process may collide with the
`mcmini` process or even the next branch. The race occurs when the
old target branch is alive at the time that `mcmini` decides to clear
the semaphores in shared memory for the next branch.

The implementation sends a `kill(branch_pid, SIGUSR1)` upon destruction of
`local_linux_process`. However, `kill(3)` is asynchronous, so the
process that is being targeted may not be killed once `mcmini` resumes
execution. To maintain the invariant that we wait for that process to be
killed before continuing execution, we simply call `wait()` at process
destruction time in the destructor of `local_linux_process`.