Add working multithreaded_fork() under DMTCP
There was a subtle bug in the multithreaded_fork() code which only
appeared when the code was executed under DMTCP at restart-time. The
trouble was that the functions `getTLSPointer()` and `setTLSPointer()`
were being linked dynamically instead of statically. However, the
dynamic linker uses thread-local storage when performing its symbol
lookup. This causes a crash since DMTCP moves the thread local storage
under our feet. The point of setTLSPointer() and getTLSPointer is
precisely to adjust the TLS pointer for the _new_ child thread (using
the TLS of the parent is now different).

The solution is to mark the functions as `static` to ensure that they
are linked statically instead of dynamically. In this way, the TLS
pointer is avoided until it is properly adjusted by the two functions.