Add final impl of multithreaded_fork() with McMini
This commit resolves the final (known) bug with McMini and
multithreaded_fork(). The bug arose because of the protocol that McMini
used to communicate with the template process and the template process'
subsequent children.

Upon creating a new branch process, McMini would wait for the template
to write to shared memory the process ID of the new branch. This process
ID is needed later when McMini decides to explore a new branch and needs
to clean up the old one via kill(3). Since the branch is a grandchild of
the McMini process (the branch is a child of the template process), the
process ID is not immediately accessible via e.g. the return call of
fork(3), hence why the template process must give the ID to McMini.

The trouble arises because in the case of multithreaded_fork(), the
process that is created by `dmtcp_restart` plays two different roles
depending on how the deep debugging is to proceed. If we are using the
"naive" version of deep debugging, i.e. without multithreaded_fork() and
instead via repeatedly calling `dmtcp_restart`, then the creation of the
process serves as the signal for McMini that the branch is ready.

In the case of multithreaded_fork(), McMini used the
`fork_process_source` construct. When prompted to create a new branch
process, McMini would wait for the template to signal the new branch was
ready for execution before continuing with model-checking. The trouble
is that the `fork_process_source` is designed for classic model checking
starting from the beginning of program execution. This means that McMini
wouldn't wait _for the creation of the template itself_. Since the
template process is the same as that used in "naive" deep debugging,
we'd erroneously signal McMini after the template was finished, but this
would be interpreted as "a new branch has been created" by the
`fork_process_source()` and we'd have a race between the template and
the McMini process for `cpid`.

The solution is to have McMini consume the first signal inside a
new `multithreaded_process_source` AND for the _first_ branch process to
perform the state transfer when opening the FIFO. The trouble is that if
we performed the state transfer BEFORE entering the template loop, we'd
be in a deadlock as McMini would be waiting for the template process to
signal it while the template process would be waiting for McMini to
open the FIFO (opening a FIFO blocks the writer until a reader is
available).