Begin adding fixes to multithreaded fork
There's a subtle bug involving multiple consecutive calls to
multithreaded_fork() that was introduced when we enabled SIGCHLD
callbacks inside the template process to detect abnormal process
termination at runtime. This commit adds a change to resolve a race
condition that occurred whenever the template thread was restored by
`dmtcp_restart` BEFORE a userspace thread. The problem was that the
`threadIdx` would not be incremented by the template thread. This meant
that a userspace thread could compute the same index as the template
thread which wrote a `0` into `origTid` used later to indicate that the
given thread should be skipped for a `clone(2)` call.