Remove redundant `sem_wait(3)` call from multithreaded_fork()
`sem_wait()` isn't async-signal-safe, so calling it within a signal
handler may cause unexpected bugs. Such an unexpected bug was observed
in prior commits: receiving a signal when calling `sem_wait` inside of
the signal handler caused unexpected behavior.

Fortunately, we can avoid calling `sem_wait()` in the handler. The key
idea is that the parent thread calling `multithreaded_fork()` is only
waiting for the child threads to call `getcontext()`. Once this happens,
the parent thread can simply call `fork()`. The cloned parent thread in
the child process will use the contexts saved from before to restore the
userspace threads. Consequently, the `sem_wait()` can be removed.

Even better would be to use the deep debugging system along
`multithreaded_fork()` and avoid the signalling entirely. This is the
work of subsequent commits