Avoid race b/w read/write shm_global_variable
Added a function to write shm_global_variable address which is protected
by the mutexes. Another function to read shm_global_variable with a
thread local variable to make sure it read only once. This way we avoid
the race happening between reading and writing of this shared global
variable.