FIX: Lower default maxTotalDepth for livelock
Previously, if the -M flag was not specified, McMini would crash with
a segfault while checking for livelock, since maxTotalDepth defaulted
to the size of the transition stack. Extending the trace caused a write
beyond the stack bounds.
This commit fixes the issue by setting the default to a lower value when
the -l flag is passed.